decimal gebruiken bij distance ipv int voorbeeld route + 10

+++ taxi_tilburg_backend/Database/Models/LocationConnection.cs
@@ -5,7 +5,7 @@ public class LocationConnection
     public required Location StartingPoint { get; set; }
     public int StartingPointId { get; set; }
     public int TravelTimeInSeconds { get; set; }
-    public int DistanceInKilometers { get; set; }
+    public decimal DistanceInKilometers { get; set; }
     public required Location EndPoint { get; set; }
     public int EndPointId { get; set; }
 }
