+++ taxi_tilburg_backend/Models/Responses/TaxiStop.cs
@@ -3,6 +3,6 @@ namespace taxi_tilburg_backend.Models.Responses;
 public class TaxiStop
 {
     public int Id { get; set; }
-    public int TotalDistance { get; set; }
+    public decimal TotalDistance { get; set; }
     public int TotalTravelTime { get; set; }
 }
